requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
gradio>=4.0.0
pypdf>=3.0.0
requests>=2.31.0
pydantic>=2.0.0

# Logging and utilities
loguru>=0.7.0